+++ src/lib.rs
@@ -28,6 +28,7 @@
 //! }
 //! ```
 
+pub mod audit;
 pub mod authenticator;
 pub mod commands;
 pub mod http;
@@ -35,6 +36,7 @@ pub mod http_client;
 pub mod jwt;
 pub mod key;
 pub mod parameters;
+pub mod rate_limit;
 pub mod system_identity;
 pub mod token;
 pub mod token_retriever;